Day 2
Today, I worked on figuring out problem 3. In the beginning, I was able to find a way but it was very inefficient. So, after a retrial, I was able to make it much more simple by using two while loops. For problem 4, I am still trying to see how I can start it. I didn't really think of doubling while loops because we did not really do that before but now I know that it is possible.