Day 3
I restarted this problem and I have a lot of trouble with this.  I think I figured out how to multiple every number within the range of three-digit numbers. However, line 12 is out of range for some reason. I need to figure out how to analyze the individual numbers and also how to adapt the code depending on the length of the product. Potentially, if I make the product digits individualized, I can break them apart like a list and see. But I wonder how long each runs will take.