Day 4
For my day 4, I had a lot of trouble with the last part of my code since last time. I tried many different ways and rewrote my code. However, by researching how to flip a number, I found a solution. From there, it was quite easy because a number like 6006 had to be equal even when it was flipped. I'm really happy I got this done and I also started to take a look at problem 5. I think problem 5 will be pretty easy after making problem 4 because the fundamental idea is the same. We filter everything by trying to divide it with every number. After doing so, we see if the new number is lower than the previous.