Day 5
I thought this one was extra fun because I thought I knew what I was doing at the beginning but It turns out that I didn't.  At the beginning, I wanted to make an efficient program by using a for loop. However, I had some trouble with how to stop running the for loop. Thus, I thought about it and realized that a simple greater than symbol would do the trick. Although I knew it isn't an efficient program, it is not as different from the one I was trying to make. At the end, I ended up making the simplest algorithm. It took about 50 seconds to come up with 232792560. How could I make this more efficient?