Day 6
In this work session, I was able to complete problems 5 and 6 pretty quickly. Question 6 was quite easy because it used the for loop function just like problems 5 and 4. We just figure out that we need to run through the range of 1-100 and then square them. For problem 7, it was even easier as we just had to figure out the formula of prime numbers. After doing so, I optimized it to get the result we wanted. This should work for every number greater than the 13th prime.