Day 7
In the beginning, I frankly did not know what adjacent numbers meant in the context of the box. But after realizing that it is just an indented string of numbers, I figured out what I was supposed to do. Fundamentally, I used for loops to go through every number until the 13th and then seeing if it is bigger than the previous number. It goes through the list by increasing i by one everytime, calculating every possibility. It is a simple algorithm but not the most efficient because it has to calculate by every shift. I struggled a little with how to get the 13 numbers to multiply each other and got confused on how I would shift the numbers to the right after every trial. But I realized it is simply just a for loop.